requirements.txt
# DR-TB Prediction Web Interface Requirements

# Web Framework
streamlit>=1.28.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.8.0

# Machine Learning
scikit-learn>=1.0.0
imbalanced-learn>=0.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Interpretability
grad-cam>=1.5.0

# Bioinformatics
biopython>=1.79

# Web Scraping
beautifulsoup4>=4.10.0

# Excel Processing
openpyxl>=3.0.0

# Progress Bars
tqdm>=4.64.0

# Utilities
pathlib2>=2.3.7; python_version < '3.4'

